Remove g:mucomplete#user_mappings
I don't need it. I will add my completion methods directly to the
dictionary s:compl_mappings.

The definition of the variable follows:

A Dictionary of user-defined completion methods. For example:

  let g:mucomplete#user_mappings = {
      \ 'sqla' : "\<c-c>a"
      \ }

Each key in the dictionary is an arbitrary string providing the name
of the completion method (but you cannot override the predefined
names) and the corresponding value is the mapping that triggers the
desired completion (see |sql-completion-static| for the definition
of <c-c>a).

User-defined completion methods may be used as predefined methods,
e.g.:

  let g:mucomplete#chains = {}
  let g:mucomplete#chains.sql = ['sqla', 'keyp', 'tags']