Remove g:mucomplete#exit_ctrlx_keys
Here's what the plugin says about it:

    The key combo to exit CTRL-X submode. You should never need to
    change this value, unless you have remapped CTRL-B in Insert mode.

    let g:mucomplete#exit_ctrlx_keys = "\<c-b>\<bs>"

For the moment, we'll keep the default value.
It could cause some problems since we have remapped C-B to move the
cursor back in insert mode.
But we don't have a lot of choice, in insert mode, all control
keys seem to be already taken.

Maybe we'll have to temporarily unmap and remap the C-B key each time we
complete something.

We'll see how it goes without doing anything…