Remove the extensions of g:mucomplete#can_complete
The variable contains a dictionary whose keys are filetypes, and values
are dictionaries of methods.

In those dictionaries, the value of each key (method) is a condition to
check for the method to be applied.
The old code was adding default values to whatever the user had defined
in their vimrc.

If I need to tweak the conditions, I will do it in the plugin directly.
So, I don't need the extensions.